feat: Adaptively tune the coefficients of loss terms.
The coefficients of loss term will evolve according to the average loss. This will split training process into 3 stages. 1: initial shaping stage. Pushing admissible set to be positive, and inadmissible set to be negative. 2. shape CBF with decent condition. 3. Decrease the coefficients of performance loss and focus on safety loss.